pyproject.toml
[tool.poetry]
name = "kwbot"
version = "0.1.0"
description = "replies with a random action to commands (non-deterministic)"
authors = ["OpenAlcoholics <team@openalcoholics.group>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.0"
requests = "^2.28.1"
geonamescache = "^1.5.0"
cinemagoer = "^2022.12.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
